Abhishek/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
requests
google-generativeai
python-dotenv
python-multipart
tqdm
